CRUD Region Working !

+++ Project/OqFazer/src/br/com/fatec/oqfazer/core/service/RegionServiceImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import com.google.common.collect.Lists;
 
 import br.com.fatec.oqfazer.api.dao.CityDAO;
+import br.com.fatec.oqfazer.api.dao.EventDAO;
 import br.com.fatec.oqfazer.api.dao.RegionDAO;
 import br.com.fatec.oqfazer.api.dto.CityDTO;
 import br.com.fatec.oqfazer.api.dto.RegionDTO;
@@ -19,11 +20,13 @@ public class RegionServiceImpl implements RegionService, CityService {
 	
 	private RegionDAO regionDAO;
 	private CityDAO cityDAO;
+	private EventDAO eventDAO;
 	private RegionDTOConverter regionDTOConverter;
 	
 	public RegionServiceImpl() {
 		this.regionDAO = ImplFinder.getImpl(RegionDAO.class);
 		this.cityDAO = ImplFinder.getImpl(CityDAO.class);
+		this.eventDAO = ImplFinder.getImpl(EventDAO.class);
 		this.regionDTOConverter = ImplFinder.getFinalImpl(RegionDTOConverter.class);
 	}
 	
@@ -31,7 +34,14 @@ public class RegionServiceImpl implements RegionService, CityService {
 	public RegionDTO insert(RegionDTO regionDTO){
 		Region region = this.regionDTOConverter.toEntity(regionDTO);
 		Long id = this.regionDAO.insertRegion(region);
-		this.cityDAO.insertCity(id, this.regionDTOConverter.toEntityCity(regionDTO.getCities()));
+		try{
+			this.cityDAO.insertCity(id, this.regionDTOConverter.toEntityCity(regionDTO.getCities()));
+		}catch(Exception e){
+			String[] a = e.getMessage().split(" ");
+			if(a[2].equals("constraint")){
+				regionDTO.setErro("Não foi possível salvar esta região, pois contem cidades ja cadastradas");
+			}
+		}
 		regionDTO.setId(id);
 		return regionDTO;
 	}
@@ -45,8 +55,14 @@ public class RegionServiceImpl implements RegionService, CityService {
 
 	@Override
 	public void delete(Long idRegionDTO) {
-		this.cityDAO.deleteCity(idRegionDTO);
-		this.regionDAO.deleteRegion(idRegionDTO);
+		if(this.eventDAO.searchEvenstsByIdRegion(idRegionDTO).isEmpty()){
+			this.cityDAO.deleteCity(idRegionDTO);
+			try{
+				this.regionDAO.deleteRegion(idRegionDTO);
+			}catch(Exception e){
+				System.out.println(e.getMessage());
+			}
+		}
 	}
 	
 	@Override
@@ -61,6 +77,7 @@ public class RegionServiceImpl implements RegionService, CityService {
 	}
 
 	@Override
+	
 	public RegionDTO searchById(Long idRegionDTO) {
 		return this.regionDTOConverter.toDTO(this.regionDAO.searchRegionById(idRegionDTO));
 	}
